+++ webview/vite.config.ts
@@ -1,0 +1,19 @@
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react";
+import cssInjectedByJsPlugin from "vite-plugin-css-injected-by-js";
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  plugins: [react(), cssInjectedByJsPlugin()],
+  build: {
+    emptyOutDir: true,
+    rollupOptions: {
+      output: {
+        entryFileNames: "assets/index.js",
+        chunkFileNames: "assets/[name].js",
+        assetFileNames: "assets/[name][extname]",
+        dir: "../out/webview",
+      },
+    },
+  },
+});