+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
-- No unreleased updates.
+### Changed
+
+- Change the result view to use WebView instead of the Output tab.
 
 ## [0.0.1] - 2023-03-06